tests: common module
move fixtures to a common module - these will be reusable after all.

changelog: none
